V0: deal_player_hands() now functional
Had to rollback the currentDealer functionality, if you can call it that, to get the deal_player_hands() to work.  Storing pointers to vectors in temp variables was giving me some problems because I think it was calling a copy.  Even though move_one_card() was pulling out a tempCard from source, setting tempVector to *destination and inserting into tempVector wasn't working.  Was it calling a copy constructor under the hood?  Regarldess, some CLI debugging found the source of the bug and I fixed it by reverting that statement back to (*destination).push_back(tempCard);  All is well.  Next task, interface?