V0:  Working on printing melds
While writing show_all_melds(), and the functions it calls, I felt it necessary to prototype a TIP function to print a meld.  The prototype was written into the class but the function hasn't been defined yet.  Write print a meld and then go back to defining show_all_melds() and the two functions it calls in turn.