V0: Incremental commit.
Continue working on printing melds with meld numbers to enable the user to select a meld.  Also, consider refactoring the entire process to return a container that holds copies of all the eligible melds.  This may necessitate a new member variable to identify a single card to a meld which would also allow proper printing of that meld, not to mention storage.  SPOILER ALERT: The "exposed" melds will be stored in a shared_ptr<vector<shared_ptr<PCard>>> exposedMelds.