Version_0: Extricated sorting/printing meld code
Extricated dupe code for finding melds into it's own function.  Lots of clean up on this commit.  Removed some methods.  Added methods.  The short version is that now a Player's Hand will add vectors of melds to a private member and then print them when it's time.  I discovered a bug in which the temp vector wasn't being cleared but I think I fixed it.  If there are two non-sequential melds in the same suit then you've replicated the input for this bug.