Version_0: Fixed sneaky 'copy constructor' bug
In T_I_G::score_the_game(), it is determined if each player is burned or not.  If they are, their got_burned() public method is called.  I'm guessing that the situation in which I was iterating through the vector of players for (auto player : players) was calling the copy constructor because the value wasn't being permanently changed by the public method.  After many debug statements and a change in s_t_g()'s prototype, I discovered and fixed the bug.  I (belive I) removed all reference to the change in s_t_g()'s prototype but I merely commented out the DEBUGGING statements.