Version_0: Refactored TIP card printer
Refactored Tong_Its_Player card printers to be more modular.  Instead of dedicated/hard-coded to print the playersHand, they now take a paramter of the same data type of playersHand while also taking a parameter to print card numbers or not.  Now that this is all done and it compiles... print the exposed melds horizontally.