Version_0: Refactored T_I_G::User Interface
Tong_Its_Game user interface now utilizes a temporary variable to store a card drawn from the deck in order to print details about the rank and suit to the user after the game status is reprinted after the draw is resolved.  Almost used the TIG card printer but that was too confusing so I just wrote a text string instead.