Version_0: Futzing w/ TIAP inheritance
The TIG players vector is "Object Slicing" Tong Its AI Player members and methods.  In an attempt to circumvent this (since a lot of the game structure revolves around iterating over the game's players) I'm going to transition players from a vector of TIP to a vector of unique_ptrs to TIPs.  This commit is a 'rollback' commit (if even such a thing exists) since a lot of find/replace is going to be involved in transitioning players.