Version_0: Huge refactor complete
TIG vector of players was "Object Slicing" the new AI derived class.  Solution was to modify the players vector to be smart pointers and *let* it object slice the TIAP players.  The TIP functionality doesn't need any of the AI functionality.  I instantiated the TIAP objects with shared pointers, pushed the shared pointers on to the vector but kept a separate copy of the AI player shared pointers for the sake of TIG calling their user interfaces.

Doing so meant a project-wide modification from vector of objects to a vector of object pointers.  Everything had to be modified.  periods became arrows, references became pointers, etc.  It got nasty but it appears to be complete.

NOTE: There might be a problem when it comes time to build "challenge" decision algorithms into the TIAP.  "Let us see." -Mrs. Harkleroad